signup email functional and debugged
