+++ server/models/mailer.js
@@ -1,13 +1,13 @@
-const db = require('../db')
-const Model = require('./model-helper')
+// const db = require('../db')
+// const Model = require('./model-helper')
 
-const Mailer;
-module.exports = Mailer;
+// const Mailer;
+// module.exports = Mailer;
 
 
-/* SEND EMAIL MESSAGE
+// /* SEND EMAIL MESSAGE
   
-*/
-Mailer.sendConfirmation = function(mailObj) {
-
-}
+// */
+// Mailer.mailReadyObj = function() {
+  
+// }
